Delete patients for practice was not working
Added some code in order to fix the action "remove_patients". It was NOT removing the patients (qdm_patients). I also added some code in order to remove the individual results for a patient. I'm not sure if that is correct.

Please review this change, although it is working for me, maybe there is an optimal way of doing it.
Thanks